Optimizer for Solar Clear sky

Find the best parameter delta for fitting clear sky radiation.

clearsky_delta(x, Ct, lower = 0, upper = 3, by = 0.01, ntol = 0)

Arguments

x

Numeric vector, time series of solar radiation.

Ct

Numeric vector, time series of clear sky radiation.

lower

Numeric scalar, lower bound for grid of delta parameters used for optimization. Default is 0.

upper

Numeric scalar, upper bound for grid of delta parameters used for optimization. Default is 3.

by

Numeric scalar, step for the grid. Default is 0.01.

ntol

Integer scalar, Tolerance for the maximum number of violations admitted of the condition Ct > x. The default is 0.

Details

Detect the best parameter delta such that the condition x > delta * Ct is satisfied for all the observations except for ntol observations.

Note

Version 1.0.1