Fit the seasonal model for clear sky radiation.

Fit the seasonal model for clear sky radiation.

clearsky_optimizer(
  x,
  dates,
  clearsky,
  lat,
  spec = seasonalClearsky_spec(),
  control = control_seasonalClearsky(),
  seasonal_model_Ct
)

Arguments

x

Numeric vector, time series of solar radiation.

dates

Character or Date vector, time series of dates.

clearsky

Numeric vector, time series of target clear sky radiation.

lat

Numeric scalar, reference latitude. Overwritten when seasonal_model_Ct is specified.

spec

See the function seasonalClearsky_spec for more details. Overwritten when seasonal_model_Ct is specified.

control

See the function control_seasonalClearsky for more details. Overwritten when seasonal_model_Ct is specified.

seasonal_model_Ct

Optional object of the class seasonalClearsky. When submitted lat, spec and control parameters used are recovered from the specified model.

Note

Version 1.0.0

Examples

x  = solarr::spec$data$GHI
dates  = solarr::spec$data$date
lat = solarr::spec$coords$lat
clearsky = solarr::spec$data$clearsky
# Without a clearsky model
seasonal_model_Ct <- clearsky_optimizer(x, dates, clearsky, lat)
# With a clearsky model
clearsky_optimizer(x, dates, clearsky, seasonal_model_Ct = seasonal_model_Ct)
#> -------------------------- seasonalClearsky --------------------------
#> Error in str_split(msg_0, ""): could not find function "str_split"