Control parameters for a solar option

Control parameters for a solar option

control_solarOption(
  nyears = c(2005, 2025),
  K = 0,
  leap_year = FALSE,
  nsim = 200,
  ci = 0.05,
  seed = 1
)

Arguments

nyears

numeric vector. Interval of years considered. The first element will be the minimum and the second the maximum years used in the computation of the fair payoff.

K

numeric, level for the strike with respect to the seasonal mean. The seasonal mean is multiplied by exp(K).

leap_year

logical, when FALSE, the default, the year will be considered of 365 days, otherwise 366.

nsim

integer, number of simulations used to bootstrap the premium's bounds. See solarOption_historical_bootstrap.

ci

numeric, confidence interval for bootstrapping. See solarOption_historical_bootstrap.

seed

integer, random seed for reproducibility. See solarOption_historical_bootstrap.

Note

Version 1.0.0.

Examples

control_options <- control_solarOption()