Seasonal clear-sky model

R6 implementation of a clear-sky seasonal model for solar radiation, using extraterrestrial radiation and seasonal harmonics.

Note

Version 1.0.3

Super class

solarr::seasonalModel -> seasonalClearsky

Public fields

lat

Numeric, scalar, latitude of the location considered.

spec

List, model specification

Active bindings

ssf

See seasonalSolarFunctions for more details.

Methods

Public methods

seasonalClearsky$new()

seasonalClearsky$fit()

seasonalClearsky$predict()

seasonalClearsky$differential()

seasonalClearsky$print()

seasonalClearsky$clone()

Inherited methods

solarr::seasonalModel$update()

solarr::seasonalModel$update_coefs_names()

solarr::seasonalModel$update_std.errors()

Method new()

Initialize a seasonalClearsky object.

Usage

seasonalClearsky$new(
  spec = seasonalClearsky_spec(),
  control = control_seasonalClearsky()
)

Arguments

spec

Named list, model's specification. See the function seasonalClearsky_spec for more details.

control

Named list, control parameters. See the function control_seasonalClearsky for more details.

Method fit()

Fit the seasonal model for clear sky radiation.

Usage

seasonalClearsky$fit(x, dates, lat, clearsky)

Arguments

x

Numeric vector, time series of solar radiation.

dates

Character or Date vector, time series of dates.

lat

Numeric scalar, reference latitude.

clearsky

Numeric vector, time series of target clear sky radiation.

Method predict()

Predict method for seasonalClearsky object.

Usage

seasonalClearsky$predict(n, newdata)

Arguments

n

Integer, scalar or vector. number of day of the year.

newdata

An optional data frame in which to look for variables with which to predict. If omitted, the fitted values are used.

Method differential()

Differential method for seasonalClearsky object.

Usage

seasonalClearsky$differential(n, newdata)

Arguments

n

Integer, scalar or vector. number of day of the year.

newdata

An optional data frame in which to look for variables with which to predict. If omitted, the fitted values are used.

Method print()

Print method for seasonalClearsky object.

Usage

seasonalClearsky$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

seasonalClearsky$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.