Specification for seasonalClearsky model

Specification for seasonalClearsky model

seasonalClearsky_spec(
  method = "OLS",
  order = 1,
  order_H0 = 1,
  method_H0 = "spencer",
  period = 365,
  include.intercept = TRUE,
  include.trend = FALSE
)

Arguments

method

Character, method of fitting: can be: OLS, RLS, CLS

order

Integer scalar, number of combinations of sines and cosines.

order_H0

Integer scalar, order of the polynomial for extra-terrestrial radiation.

method_H0

Character, method of computation of the extra-terrestrial radiation: can be: spencer, cooper.

period

Integer scalar, seasonality period. The default is 365.

include.intercept

Logical, when TRUE, the default, the intercept will be included in the clear sky model.

Value

Named list of control parameters.

Note

Version 1.0.0

Examples

spec_Ct = spec_seasonalClearsky()
#> Error in spec_seasonalClearsky(): could not find function "spec_seasonalClearsky"