Distribution test

Evaluate a Kolmogorov-Smirnov test on the residuals of a solarModel model object against the estimated Gaussian mixture distribution.

solarModel_test_distribution(
  model,
  H0 = "gm",
  type = "full",
  ci = 0.05,
  min_quantile = 0.025,
  max_quantile = 0.985
)

Arguments

model

An object of the class solarModel

H0

Character, null hypothesis for the residuals distribution. Can be:

"gm"

for a test against a gaussian mixture distribution;

"norm"

for a test against a gaussian distribution.

type

character(1), type of test. Can be:

"train"

the test is performed only on train data;

"test"

the test is performed only on test data;

"full"

the test is performed only on the full data.

ci

p.value for rejection.

min_quantile

minimum quantile for the grid of values.

max_quantile

maximum quantile for the grid of values.

Note

Version 1.0.0.

Examples

model = solarModel$new(spec)
model$fit()
#> No outliers!
solarModel_test_distribution(model)
#> # A tibble: 12 × 9
#>    Month H0    type      n alpha    KS rejection_lev p.value result
#>    <int> <chr> <chr> <int> <chr> <dbl>         <dbl>   <dbl> <chr> 
#>  1     1 gm    full    651 5%    0.858          1.36   0.453 Passed
#>  2     2 gm    full    593 5%    0.708          1.36   0.698 Passed
#>  3     3 gm    full    651 5%    0.806          1.36   0.535 Passed
#>  4     4 gm    full    630 5%    0.759          1.36   0.612 Passed
#>  5     5 gm    full    651 5%    0.677          1.36   0.749 Passed
#>  6     6 gm    full    609 5%    0.573          1.36   0.898 Passed
#>  7     7 gm    full    620 5%    0.773          1.36   0.589 Passed
#>  8     8 gm    full    620 5%    0.911          1.36   0.377 Passed
#>  9     9 gm    full    600 5%    0.871          1.36   0.434 Passed
#> 10    10 gm    full    620 5%    0.541          1.36   0.932 Passed
#> 11    11 gm    full    600 5%    0.551          1.36   0.922 Passed
#> 12    12 gm    full    620 5%    0.602          1.36   0.861 Passed