Compute the first fourth moments and variance of a SoRad

Compute the first fourth moments and variance of a SoRad

solarOption_moments(
  moments,
  transform,
  lambda = 0,
  put = TRUE,
  link = "invgumbel",
  control_options = control_solarOption()
)

Arguments

transform

slot transform of a solarModel

lambda

Sugeno parameter

put

Logical, when TRUE, the default, will be computed the price for a put contract, otherwise for a call contract.

control_options

Named list with control parameters. See control_solarOption for more details.

Note

Version 1.0.0.

Examples

# Solar model
model <- solarModel$new(spec)
model$fit()
#> No outliers!
moments <- model$moments$unconditional
solarOption_moments(moments[1:365,], model$transform)
#> Error in 1:nrow(df_tT): argument of length 0